Breaking examples into their own directory.  Fixed README.
