+++ examples/TextScroll/funconfig.h
@@ -1,0 +1,11 @@
+#ifndef _FUNCONFIG_H
+#define _FUNCONFIG_H
+
+#define RANDOM_STRENGTH          2
+#define CH32V003                 1
+
+#define FUNCONF_USE_DEBUGPRINTF  1
+#define FUNCONF_USE_UARTPRINTF   0
+
+#endif
+
